examples/sitesage/pyproject.toml
[project]
name = "sitesage"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chainless>=0.2.0",
    "playwright>=1.52.0",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
]

